dramatically improve magic paint performance on large canvases
By drawing canvas to a fixed viewport instead of the page it eliminates potentially crashing performance issues with bigger spaces.

The downside of this approach is that a paint stroke remains in the same place as you scroll, instead of scrolling away like it used to.

(drawing a circle in a 1,000 x 1,000px canvas is dramatically faster than drawing that same circle in a 10,000 x 10,000px canvas)

I'm not sure if anyone will mind this irl tho.