remove cards from dom when not isvisibleinviewport
Don’t render card at all if:

x Is out of viewport (not just hiding media)

On a resize w 500 cards, the matrix is applying to all, per frame. Whether the image is in the card doesn’t seem to affect things. Take the card off the dom completly

x Connection path should not update if one of the cards is not found