reduce header ui btns
2 related UI buttoin experiments I'm interested in trying out:

hide the header comment button to see if anyone uses/misses it
in the footer, replace the user settings shortcut button with the presentation mode toggle button. I think presentation mode makes more contextual sense next to zoom, and I suspect user settings  isn't a thing that's opened frequently enough to be worth such prominent real estate.

the net result of these changes are a more cleaner and focused header. Also I want to increase the likelihood of ppl eventually discovering the sidebar options, and so a lower ui intimidation/confusion factor increases the likelihood of that