fix copy err
https://forum.kinopio.club/t/copy-to-space-bug-critical-failure/1608/6

when copying multiple items which shared the same connection type,
the connections weren't being assigned to the correct new connection type

causing sql errs on server
and client side errs on load when looking up the connectionType

🚛 fixed the copy/move code to update connections to the correct new type id

this should now be resolved. Thanks so much for your help debugging this!