set lower timeout, app goes into 'offline' mode when server is slow/unresponsive
basically there are 3 loading scenarios:
1. server is offline : client skips over all remote data requests and the app works offline
2. server is online and running fine: everything works
3. server is online, but super slow or not responding to requests: client gets stuck waiting for the server and that's why you see a blank screen

short term solution: i have a beta that reduces the server request timeout time from 40s to 5s. And if the server takes longer than 5s , then the app will load in offline mode.

So the longest you'll see a white screen in scenario #3 will be 5s , and after that you'll at least be able to access your offline data and make changes that'll be synced when the server is fully back